entropy_estimators: move discretization of discrete-variable vectors into hist
We need to handle number vectors from discrete random variables as single
outcomes, so we "discretize" them by turning them into tuples.  We formerly did
that on a per-column basis; now we do it in hist() so we don't have to care
about it everywhere else.

Testing Done: calculate TC(X) and TC(X;Y) correctly

Signed-off-by: Eli Sennesh <eligottlieb@gmail.com>